add local file md5 after file transmit
LOCAL_TO_S3 mode:
Before this fix, it calculate the local file md5 while transmission. But if the source file is changing while transmission, file might not be found err, but it is actually broken. So now if enable ifVerifyMD5=True, after transmit every file, it will re-read the source file from local and calculate MD5 again to avoid above case. It will slow down performance, this feature is not enable by default.